+++ RELEASE_NOTES.md
@@ -16,6 +16,10 @@ v1.0.0-beta.x.x
   `TraitsData`.
   [#1153](https://github.com/OpenAssetIO/OpenAssetIO/issues/1153)
 
+- Modified no-argument constructor of `Context` to create an empty (but
+  non-null) `locale`.
+  [#1153](https://github.com/OpenAssetIO/OpenAssetIO/issues/1153)
+
 v1.0.0-beta.1.0
 ---------------